Clean up commit message - Remove references to deleted files - Fix handling of global_stats in plotting.py - Removed commented out modules and code - Fixed imports to use sbgm.variable_utils instead of data_analysis_pipeline.stats_analysis.variable_utils - Fixed imports - Removed unused imports
