+++ sbgm/data/datasets_loading.py
@@ -9,7 +9,7 @@ from glob import glob
 from datetime import datetime
 
 from data_analysis_pipeline.stats_analysis.path_utils import build_data_path
-from data_analysis_pipeline.stats_analysis.variable_utils import crop_to_region, get_var_name_short, correct_variable_units
+from sbgm.variable_utils import crop_to_region, get_var_name_short, correct_variable_units
 from concurrent.futures import ProcessPoolExecutor, as_completed
 
 # Setup logging
