+++ sbgm/evaluate_sbgm/generation_main.py
@@ -6,7 +6,7 @@ from datetime import datetime
 from omegaconf import OmegaConf
 
 from sbgm.training_utils import get_model, get_gen_dataloader
-from sbgm.special_transforms import build_back_transforms, build_back_transforms_from_stats
+from sbgm.special_transforms import build_back_transforms_from_stats
 from sbgm.evaluate_sbgm.generation import SampleGenerator #run_generation_multiple, run_generation_single, run_generation_repeated
 from sbgm.utils import get_model_string
