+++ sbgm/monitoring.py
@@ -1,7 +1,25 @@
+"""
+    Monitoring functions for EDM.
+    Currently includes:
+        - edm_cosine_metric: Cosine similarity metric for EDM models as per Karras et al. (2022).
+        - _masked_corrcoef_per_sample: Mean Pearson correlation across the batch, computed per-sample over masked pixels.
+        - report_precip_extremes: Reports extremes in a back-transformed precipitation tensor.
+
+    TODO:
+        - FSS (Fractions Skill Score) implementation at 5, 10, 20 km scales
+        - PSD slope metric
+        - Q95/Q99 metrics
+        - Wet-day frequency 
+        - Other metrics from "Evaluating Generative Models via Precision and Recall" (Sajjadi et al. 2018)
+"""
+
 import torch
 import logging
+import os
 
 import torch.nn.functional as F
+from datetime import datetime
+
 from sbgm.losses import EDMLoss
 
 logger = logging.getLogger(__name__)
@@ -105,3 +123,35 @@ def report_precip_extremes(x_bt: torch.Tensor, name: str, cap_mm_day: float = 50
 
     return {'has_extreme': False}
 
+
+def setup_logger(log_dir, name="train_log", log_to_stdout=True):
+    # Set up the path for the log directory
+    os.makedirs(log_dir, exist_ok=True)
+    timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
+    log_path = os.path.join(log_dir, f"{name}_{timestamp}.log")
+
+    # Set up a logger, with level set to INFO which means it will log INFO, WARNING, ERROR, and CRITICAL messages
+    logger = logging.getLogger()
+    logger.setLevel(logging.INFO)
+
+    # Remove existing handlers (we remove all handlers to avoid duplicates)
+    for handler in logger.handlers[:]:
+        logger.removeHandler(handler)
+
+    # File handler to write logs to a file
+    file_handler = logging.FileHandler(log_path)
+    file_handler.setLevel(logging.INFO)
+    # Set the format for the log messages
+    file_formatter = logging.Formatter('%(asctime)s | %(levelname)s | %(message)s')
+    # Apply the formatter to the file handler
+    logger.addHandler(file_handler)
+
+    # Optional: also print to terminal
+    if log_to_stdout:
+        stream_handler = logging.StreamHandler()
+        stream_handler.setLevel(logging.INFO)
+        stream_handler.setFormatter(file_formatter)
+        logger.addHandler(stream_handler)
+
+    logger.info(f"Logging to {log_path}")
+    return logger
